# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  xp: number;
  level: number;
  streak: number;
  xpToNextLevel: number;
  xpForCurrentLevel: number;
  /** @nullable */
  avatar?: string | null;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  avatar?: string | null;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  xpReward: number;
  priority: TaskPriority;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  xpReward?: number;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  completed?: boolean;
  priority?: TaskUpdatePriority;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type BadgeCategory = typeof BadgeCategory[keyof typeof BadgeCategory];


export const BadgeCategory = {
  streak: 'streak',
  tasks: 'tasks',
  xp: 'xp',
  level: 'level',
  subject: 'subject',
} as const;

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
  category: BadgeCategory;
}

export interface TaskUpdateResult {
  task: Task;
  xpAwarded: number;
  newBadges: Badge[];
  profile?: Profile;
}

export interface Subject {
  id: number;
  name: string;
  color: string;
  taskCount: number;
}

export interface SubjectInput {
  /** @minLength 1 */
  name: string;
  color: string;
}

export type StatsTasksPerSubjectItem = {
  subjectName: string;
  color: string;
  count: number;
};

export type StatsRecentActivityItem = {
  taskTitle: string;
  xpAwarded: number;
  completedAt: string;
};

export interface Stats {
  tasksCompletedToday: number;
  tasksCompletedThisWeek: number;
  xpEarnedToday: number;
  xpEarnedThisWeek: number;
  totalTasksCompleted: number;
  currentStreak: number;
  longestStreak: number;
  tasksPerSubject: StatsTasksPerSubjectItem[];
  recentActivity: StatsRecentActivityItem[];
}

export type GetTasksParams = {
/**
 * @nullable
 */
subjectId?: number | null;
/**
 * @nullable
 */
completed?: boolean | null;
};