# lib/api-zod/src/generated/types/badge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */
import type { BadgeCategory } from './badgeCategory';

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
  category: BadgeCategory;
}