# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  xp: number;
  level: number;
  streak: number;
  xpToNextLevel: number;
  xpForCurrentLevel: number;
  /** @nullable */
  avatar?: string | null;
}