# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsRecentActivityItem } from './statsRecentActivityItem';
import type { StatsTasksPerSubjectItem } from './statsTasksPerSubjectItem';

export interface Stats {
  tasksCompletedToday: number;
  tasksCompletedThisWeek: number;
  xpEarnedToday: number;
  xpEarnedThisWeek: number;
  totalTasksCompleted: number;
  currentStreak: number;
  longestStreak: number;
  tasksPerSubject: StatsTasksPerSubjectItem[];
  recentActivity: StatsRecentActivityItem[];
}