# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  xpReward: number;
  priority: TaskPriority;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}