# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Study XP API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  xpReward?: number;
}